added new menu items

+++ skiwm/lib/models/race_arguments.dart
@@ -1,7 +1,7 @@
 class RaceArguments {
   final String title;
   final String image;
-  final int difficulty;
+  final double difficulty;
 
   RaceArguments(this.title, this.image, this.difficulty);
 }
